# webapitc2005b/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Web API Express",
        "description": "My WebApi using Node",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3001",
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/loadUserProgress/{user_id}/{level_id}/{group_id}": {
            "get": {
                "tags": [
                    "Game"
                ],
                "summary": "Carga el progeso del jugador de acuerdo a su nivel y su grupo",
                "description": "Regresa los datos del progreso del jugador",
                "operationId": "loadUserProgress",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "user_id",
                        "description": "User ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "level_id",
                        "description": "Level ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "group_id",
                        "description": "Group ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/UserProgress"
                        }
                    }
                }
            }
        },
        "/api/updateUserProgress/{user_id}/{level_id}/{group_id}": {
            "put": {
                "tags": [
                    "Game"
                ],
                "summary": "Actualiza el progeso del jugador de acuerdo a su nivel y su grupo",
                "description": "Actualiza los datos del progreso del jugador",
                "operationId": "updateUserProgress",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "user_id",
                        "description": "User ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "level_id",
                        "description": "Level ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "group_id",
                        "description": "Group ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Datos del progreso de jugador a actualizar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserProgress"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/UserProgress"
                        }
                    }
                }
            }
        },
        "/api/loadMemoryGameProgress/{user_id}/{level_id}/{group_id}": {
            "get": {
                "tags": [
                    "Game"
                ],
                "summary": "Carga el progeso del jugador en el Memorama de acuerdo a su nivel y su grupo",
                "description": "Regresa los datos del progreso del jugador en el Memorama",
                "operationId": "loadMemoryGameProgress",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "user_id",
                        "description": "User ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "level_id",
                        "description": "Level ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "group_id",
                        "description": "Group ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/MemoryGameProgress"
                        }
                    }
                }
            }
        },
        "/api/updateMemoryGameProgress/{user_id}/{level_id}/{group_id}": {
            "put": {
                "tags": [
                    "Game"
                ],
                "summary": "Actualiza el progeso del jugador en el Memomaram de acuerdo a su nivel y su grupo",
                "description": "Actualiza los datos del progreso del jugador en el Memorama",
                "operationId": "updateMemoryGameProgress",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "user_id",
                        "description": "User ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "level_id",
                        "description": "Level ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "path",
                        "name": "group_id",
                        "description": "Group ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Datos del progreso de jugador a actualizar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MemoryGameProgress"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "schema": {
                            "$ref": "#/definitions/MemoryGameProgress"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "UserProgress": {
            "type": "object",
            "properties": {
                "minigame_1": {
                    "type": "boolean"
                },
                "minigame_1_image": {
                    "type": "string"
                },
                "minigame_1_document_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "minigame_2": {
                    "type": "boolean"
                },
                "minigame_2_image": {
                    "type": "string"
                },
                "minigame_2_document_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "minigame_3": {
                    "type": "boolean"
                },
                "minigame_3_image": {
                    "type": "string"
                },
                "minigame_3_document_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "minigame_4": {
                    "type": "boolean"
                },
                "minigame_4_image": {
                    "type": "string"
                },
                "minigame_4_document_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "minigame_5": {
                    "type": "boolean"
                },
                "minigame_5_image": {
                    "type": "string"
                },
                "minigame_5_document_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "minigame_6": {
                    "type": "boolean"
                },
                "minigame_6_image": {
                    "type": "string"
                },
                "minigame_6_document_id": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "MemoryGameProgress": {
            "type": "object",
            "properties": {
                "minigame_1": {
                    "type": "boolean"
                },
                "minigame_1_image": {
                    "type": "string"
                },
                "minigame_1_document_id": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
...
    }
}